Async handlers as a subpath exported module
