+++ package.json
@@ -25,6 +25,16 @@
         "types": "./lib/remote-data.d.ts",
         "default": "./lib/remote-data.js"
       }
+    },
+    "./async": {
+      "import": {
+        "types": "./lib/async.d.mts",
+        "default": "./lib/async.mjs"
+      },
+      "require": {
+        "types": "./lib/async.d.ts",
+        "default": "./lib/async.js"
+      }
     }
   },
   "sideEffects": false,
