+++ src/index.ts
@@ -1,16 +1,10 @@
 export {
   match,
   matchWithDefault,
-  matchAsync,
-  matchWithDefaultAsync,
-  matchAllAsync,
   map,
   mapAll,
-  mapAsync,
   fold,
   foldWithDefault,
-  foldAsync,
-  foldWithDefaultAsync,
   count,
   partition,
   createPipeHandlers,
@@ -28,9 +22,4 @@ export type {
   MapperAll,
   Folder,
   FolderWithDefault,
-  AsyncMatcher,
-  AsyncMatcherWithDefault,
-  AsyncMapper,
-  AsyncFolder,
-  AsyncFolderWithDefault,
 } from './types';
