+++ src/unions.ts
@@ -1,9 +1,4 @@
 import {
-  AsyncFolder,
-  AsyncFolderWithDefault,
-  AsyncMapper,
-  AsyncMatcher,
-  AsyncMatcherWithDefault,
   Folder,
   FolderWithDefault,
   InferUnionFromSchema,
@@ -29,7 +24,7 @@ declare global {
 
 // ── Internal helpers ───────────────────────────────────────────────────────
 
-const DEFAULT_DISCRIMINANT = 'type';
+export const DEFAULT_DISCRIMINANT = 'type';
 
 /**
  * Thrown when a discriminated union value carries a variant that no handler
@@ -67,7 +62,7 @@ function rethrow(err: Error, caller: Function): never {
   throw err;
 }
 
-function ensureUnion(
+export function ensureUnion(
   input: unknown,
   discriminant: PropertyKey,
   caller: Function,
@@ -99,7 +94,7 @@ function reduce<
   return acc;
 }
 
-function dispatch<
+export function dispatch<
   T extends SampleUnion<Discriminant>,
   Result,
   Discriminant extends PropertyKey,
@@ -122,7 +117,7 @@ function dispatch<
   return rethrow(new UnknownVariantError(key, Object.keys(handlers)), caller);
 }
 
-async function reduceAsync<
+export async function reduceAsync<
   T extends SampleUnion<Discriminant>,
   Acc,
   Discriminant extends PropertyKey,
@@ -547,209 +542,6 @@ export function partition<
 }
 
 /**
- * Async exhaustive pattern matching. Like {@link match}, but handlers may return
- * `Promise<R>` or `R`. The result is unified to `Promise<R>` — never
- * `Promise<A> | Promise<B>`.
- *
- * @example
- * ```ts
- * const profile = await matchAsync(user)({
- *   admin: async ({ id }) => fetchAdminProfile(id),
- *   guest: async ({ id }) => fetchGuestProfile(id),
- * }); // typed: AdminProfile | GuestProfile
- * ```
- */
-export function matchAsync<
-  T extends SampleUnion<Discriminant>,
-  Discriminant extends PropertyKey = 'type',
-  Payload extends any = never,
->(
-  input: T,
-  discriminant: Discriminant = DEFAULT_DISCRIMINANT as Discriminant,
-  payload?: Payload,
-): <Result>(
-  matcher: AsyncMatcher<T, Result, Discriminant, Payload>,
-) => Promise<Result> {
-  ensureUnion(input, discriminant, matchAsync);
-  return async <Result>(
-    matcher: AsyncMatcher<T, Result, Discriminant, Payload>,
-  ) =>
-    dispatch<T, Result | Promise<Result>, Discriminant, Payload>(
-      input,
-      matcher as unknown as Record<
-        string,
-        (input: any, payload: Payload) => Result | Promise<Result>
-      >,
-      discriminant,
-      undefined,
-      payload,
-      matchAsync,
-    );
-}
-
-/**
- * Async pattern matching with `Default` fallback. Like {@link matchWithDefault},
- * but handlers may return `Promise<R>` or `R`. Result unified to `Promise<R>`.
- */
-export function matchWithDefaultAsync<
-  T extends SampleUnion<Discriminant>,
-  Discriminant extends PropertyKey = 'type',
-  Payload extends any = never,
->(
-  input: T,
-  discriminant: Discriminant = DEFAULT_DISCRIMINANT as Discriminant,
-  payload?: Payload,
-): <U>(
-  matcher: AsyncMatcherWithDefault<T, U, Discriminant, Payload>,
-) => Promise<U> {
-  ensureUnion(input, discriminant, matchWithDefaultAsync);
-  return async <U>(
-    matcher: AsyncMatcherWithDefault<T, U, Discriminant, Payload>,
-  ) =>
-    dispatch<T, U | Promise<U>, Discriminant, Payload>(
-      input,
-      matcher as unknown as Record<
-        string,
-        (input: any, payload: Payload) => U | Promise<U>
-      >,
-      discriminant,
-      (matcher as any).Default,
-      payload,
-      matchWithDefaultAsync,
-    );
-}
-
-/**
- * Parallel per-item async matching across a collection. Each item is dispatched
- * to its variant handler concurrently via `Promise.all`. Returns `Promise<R[]>`.
- *
- * Use this as the parallel sibling to {@link foldAsync} when handlers are
- * independent (no shared accumulator).
- *
- * @example
- * ```ts
- * const names = await matchAllAsync(users)({
- *   admin: async ({ id }) => `admin:${await fetchName(id)}`,
- *   guest: async ({ id }) => `guest:${await fetchName(id)}`,
- * });
- * ```
- */
-export function matchAllAsync<
-  T extends SampleUnion<Discriminant>,
-  Discriminant extends PropertyKey = 'type',
-  Payload extends any = never,
->(
-  items: readonly T[],
-  discriminant: Discriminant = DEFAULT_DISCRIMINANT as Discriminant,
-  payload?: Payload,
-): <Result>(
-  matcher: AsyncMatcher<T, Result, Discriminant, Payload>,
-) => Promise<Result[]> {
-  return <Result>(matcher: AsyncMatcher<T, Result, Discriminant, Payload>) =>
-    Promise.all(
-      items.map(async (item) => {
-        ensureUnion(item, discriminant, matchAllAsync);
-        return dispatch<T, Result | Promise<Result>, Discriminant, Payload>(
-          item,
-          matcher as unknown as Record<
-            string,
-            (input: any, payload: Payload) => Result | Promise<Result>
-          >,
-          discriminant,
-          undefined,
-          payload,
-          matchAllAsync,
-        );
-      }),
-    );
-}
-
-/**
- * Async partial transform. Like {@link map}, but handlers may return a Promise.
- * Variants without a handler pass through unchanged. Result is `Promise<T>`.
- */
-export function mapAsync<
-  T extends SampleUnion<Discriminant>,
-  Discriminant extends PropertyKey = 'type',
-  Payload extends any = never,
->(
-  input: T,
-  discriminant: Discriminant = DEFAULT_DISCRIMINANT as Discriminant,
-  payload?: Payload,
-): (mapper: AsyncMapper<T, Discriminant, Payload>) => Promise<T> {
-  ensureUnion(input, discriminant, mapAsync);
-  return async (mapper) => {
-    const result = await dispatch<T, T | Promise<T>, Discriminant, Payload>(
-      input,
-      mapper as unknown as Record<
-        string,
-        (input: any, payload: Payload) => T | Promise<T>
-      >,
-      discriminant,
-      () => input,
-      payload,
-      mapAsync,
-    );
-    return result === input
-      ? result
-      : { ...result, [discriminant]: input[discriminant] };
-  };
-}
-
-/**
- * Async sequential fold. Like {@link fold}, but each handler may return
- * `Acc` or `Promise<Acc>`. The accumulator threads through `await`, so handlers
- * run strictly in array order. For parallel per-item dispatch, use
- * {@link matchAllAsync}.
- */
-export function foldAsync<
-  T extends SampleUnion<Discriminant>,
-  Acc,
-  Discriminant extends PropertyKey = 'type',
->(
-  items: readonly T[],
-  initial: Acc,
-  discriminant: Discriminant = DEFAULT_DISCRIMINANT as Discriminant,
-): (handlers: AsyncFolder<T, Acc, Discriminant>) => Promise<Acc> {
-  return (handlers) =>
-    reduceAsync(
-      items,
-      initial,
-      handlers as any,
-      discriminant,
-      undefined,
-      foldAsync,
-    );
-}
-
-/**
- * Async partial fold with `Default` fallback. Like {@link foldWithDefault},
- * but each handler (including `Default`) may return `Acc` or `Promise<Acc>`.
- * Sequential execution — accumulator threads through `await`, so handlers
- * run strictly in array order. For parallel per-item dispatch, use
- * {@link matchAllAsync}.
- */
-export function foldWithDefaultAsync<
-  T extends SampleUnion<Discriminant>,
-  Acc,
-  Discriminant extends PropertyKey = 'type',
->(
-  items: readonly T[],
-  initial: Acc,
-  discriminant: Discriminant = DEFAULT_DISCRIMINANT as Discriminant,
-): (handlers: AsyncFolderWithDefault<T, Acc, Discriminant>) => Promise<Acc> {
-  return (handlers) =>
-    reduceAsync(
-      items,
-      initial,
-      handlers as any,
-      discriminant,
-      handlers.Default,
-      foldWithDefaultAsync,
-    );
-}
-
-/**
  * Creates a pipe-friendly handler factory bound to a specific discriminant key.
  * Returns an object whose methods follow the reversed-curry shape `(handlers) => (input) => result`,
  * making them composable inside FP `pipe` utilities without wrapper lambdas.
@@ -819,37 +611,6 @@ export function createPipeHandlers<
       ...args: [Payload] extends [never] ? [] : [payload: Payload]
     ) => T,
 
-    matchAsync: wI(matchAsync) as <U, Payload extends any = never>(
-      handlers: AsyncMatcher<T, U, Discriminant, Payload>,
-    ) => (
-      input: T,
-      ...args: [Payload] extends [never] ? [] : [payload: Payload]
-    ) => Promise<U>,
-
-    matchWithDefaultAsync: wI(matchWithDefaultAsync) as <
-      U,
-      Payload extends any = never,
-    >(
-      handlers: AsyncMatcherWithDefault<T, U, Discriminant, Payload>,
-    ) => (
-      input: T,
-      ...args: [Payload] extends [never] ? [] : [payload: Payload]
-    ) => Promise<U>,
-
-    matchAllAsync: wI(matchAllAsync) as <U, Payload extends any = never>(
-      handlers: AsyncMatcher<T, U, Discriminant, Payload>,
-    ) => (
-      items: readonly T[],
-      ...args: [Payload] extends [never] ? [] : [payload: Payload]
-    ) => Promise<U[]>,
-
-    mapAsync: wI(mapAsync) as <Payload extends any = never>(
-      handlers: AsyncMapper<T, Discriminant, Payload>,
-    ) => (
-      input: T,
-      ...args: [Payload] extends [never] ? [] : [payload: Payload]
-    ) => Promise<T>,
-
     fold: wF(fold) as <Acc>(
       items: readonly T[],
       initial: Acc,
@@ -859,18 +620,6 @@ export function createPipeHandlers<
       items: readonly T[],
       initial: Acc,
     ) => (handlers: FolderWithDefault<T, Acc, Discriminant>) => Acc,
-
-    foldAsync: wF(foldAsync) as <Acc>(
-      items: readonly T[],
-      initial: Acc,
-    ) => (handlers: AsyncFolder<T, Acc, Discriminant>) => Promise<Acc>,
-
-    foldWithDefaultAsync: wF(foldWithDefaultAsync) as <Acc>(
-      items: readonly T[],
-      initial: Acc,
-    ) => (
-      handlers: AsyncFolderWithDefault<T, Acc, Discriminant>,
-    ) => Promise<Acc>,
 
     count:
       (variants: T[Discriminant] | readonly T[Discriminant][]) =>
@@ -928,7 +677,7 @@ export function createPipeHandlers<
  * ```
  */
 const RESERVED_UNION_KEYS = new Set<string>(
-  'is isKnown match matchWithDefault matchAsync matchWithDefaultAsync matchAllAsync map mapAll mapAsync fold foldWithDefault foldAsync foldWithDefaultAsync count partition variants discriminant _union'.split(
+  'is isKnown match matchWithDefault map mapAll fold foldWithDefault count partition variants discriminant _union'.split(
     ' ',
   ),
 );
